Solved #47 and fixed a lot of pausing interactions
Resuming with the B button in Bonfire or Pause menu would make player roll. if any other action was bound to B, it would be used when resuming. This has been proofed to prevent it from happening. Also, resting at Bonfire doesn't reset the player completely like it used to, preventing it from overriding the interact animation